requirements.txt
bidict==0.22.1
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
Flask==2.0.3
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.26.0
simple-websocket==1.0.0
six==1.16.0
socketIO-client==0.7.2
urllib3==1.26.18
websocket-client==1.6.4
Werkzeug==2.1.2
wsproto==1.2.0
gunicorn
eventlet